Do not use deprecated backend controller signature
Backend controller actions do not receive a prepared response
since TYPO3 v9.5 [1].

We do prepare the Response object ourselves now.
The deprecation migration suggests to use classes like
JsonResponse and HtmlResponse – we dont use these for now,
as they are only available as of TYPO3 v9. Instead we limit to use
the generic Response class which was already available in TYPO3 v8.

Also use return type declarations while we are changing the signature.

[1] https://docs.typo3.org/typo3cms/extensions/core/Changelog/9.5/Deprecation-84196-BackendControllerActionsDoNotReceivePreparedResponse.html